feat(time.Time to *timestamppb.Timestamp):

+++ internal/biz/account.go
@@ -21,6 +21,6 @@ func (au *AccountUsecase) Register(ctx context.Context, username, password strin
 func (au *AccountUsecase) Login(ctx context.Context, username, password string) (*v1.LoginResponse, error) {
 	return au.repo.Login(ctx, username, password)
 }
-func (au *AccountUsecase) RefreshToken(ctx context.Context, id uint32) (*v1.LoginResponse, error) {
+func (au *AccountUsecase) RefreshToken(ctx context.Context, id uint64) (*v1.LoginResponse, error) {
 	return au.repo.RefreshToken(ctx, id)
 }
