+++ internal/biz/repo.go
@@ -8,5 +8,5 @@ import (
 type AccountRepo interface {
 	Register(context.Context, string, string) (*v1.LoginResponse, error)
 	Login(context.Context, string, string) (*v1.LoginResponse, error)
-	RefreshToken(context.Context, uint32) (*v1.LoginResponse, error)
+	RefreshToken(context.Context, uint64) (*v1.LoginResponse, error)
 }
